.vimrc Oni support

+++ .vimrc
@@ -9,29 +9,30 @@ endif
 "Make sure to run :PlugInstall when you first run
 "vim-plug plugins
 call plug#begin('~/.vim/plugs')
-Plug 'junegunn/vim-plug'
+    Plug 'junegunn/vim-plug'
 
-Plug 'scrooloose/nerdtree'
-Plug 'Xuyuanp/nerdtree-git-plugin'
+    Plug 'junegunn/fzf', { 'dir': '~/.fzf', 'do': './install --all' }
+    Plug 'junegunn/fzf.vim'
 
-Plug 'junegunn/fzf', { 'dir': '~/.fzf', 'do': './install --all' }
-Plug 'junegunn/fzf.vim'
+    Plug 'sheerun/vim-polyglot'
 
-Plug 'sheerun/vim-polyglot'
+    Plug 'w0rp/ale'
 
-Plug 'w0rp/ale'
+    if !exists('g:gui_oni')
+        Plug 'scrooloose/nerdtree'
+        Plug 'Xuyuanp/nerdtree-git-plugin'
 
-Plug 'Valloric/YouCompleteMe', { 'do': 'npm install -g typescript'}
+        Plug 'Valloric/YouCompleteMe', { 'do': 'npm install -g typescript'}
+    endif
 
-Plug 'airblade/vim-gitgutter'
-Plug 'tpope/vim-fugitive'
+    Plug 'airblade/vim-gitgutter'
+    Plug 'tpope/vim-fugitive'
 
-Plug 'vim-airline/vim-airline'
-Plug 'vim-airline/vim-airline-themes'
+    Plug 'vim-airline/vim-airline'
+    Plug 'vim-airline/vim-airline-themes'
 
-Plug 'tpope/vim-surround'
-Plug 'tpope/vim-commentary'
-
+    Plug 'tpope/vim-surround'
+    Plug 'tpope/vim-commentary'
 call plug#end()
 
 "Settings
@@ -67,8 +68,10 @@ let mapleader=","
 let g:airline_theme='dark'
 
 "NerdTree Open when no commands specified
-autocmd StdinReadPre * let s:std_in=1
-autocmd VimEnter * if argc() == 0 && !exists("s:std_in") | NERDTree | endif
+if !exists('g:gui_oni')
+    autocmd StdinReadPre * let s:std_in=1
+    autocmd VimEnter * if argc() == 0 && !exists("s:std_in") | NERDTree | endif
+endif
 
 "Aliases
 command Ygt YcmCompleter GoTo
